Move make_malicious_contexts to reveal tests
It drops `SecurityValidator` so it is not really that useful. Reveal does not care, so it can use it
